Add excel export option for Off Road Repors

+++ database/migrations/2017_11_23_104042_create_current_parking_reports_table.php
@@ -23,7 +23,7 @@ class CreateCurrentParkingReportsTable extends Migration
             $table->double('latitude')->nullable();
             $table->double('longitude')->nullable();
             $table->double('orientation')->nullable();
-            $table->integer('speed')->nullable();
+            $table->double('speed')->nullable();
             $table->double('odometer')->nullable();
 
             /* Route report */
